requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.1
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.0.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
